cast to unsigned integers before bitshifting
bitshift on signed values is undefined in c++ standard.

http://gna.org/bugs/?25219

+++ src/map/location.cpp
@@ -57,7 +57,7 @@ const std::vector<map_location::DIRECTION> & map_location::default_dirs() {
 
 std::size_t hash_value(map_location  const & a){
 	std::hash<size_t> h;
-	return h( (a.x << 16) ^ a.y );
+	return h( (static_cast<uint32_t>(a.x) << 16) ^ static_cast<uint32_t>(a.y) );
 }
